feat: Phase P2 体验打磨 + 工程清理（4 项）
- P2-1: 批测支持选取部分用例（Checkbox 多选 + test_case_ids 后端过滤）
- P2-2: Agent 版本 Modal 字段分组（Divider）+ 裁判/模型配置查看态/编辑态切换
- P2-3: pass_threshold 类型统一为 float（消除 Decimal 序列化为字符串问题）
- P2-4: 移除 LLMAdapter.chat_json() 无用 json_schema 参数
- 审查修复: _load_context 空用例防护、CreateBatchModal 异步加载同步、auto-edit 防重复

+++ backend/app/services/judge_runner.py
@@ -1,6 +1,5 @@
 import json
 import logging
-from decimal import Decimal
 
 from app.llm.base import LLMAdapter
 from app.models.judge_config import ChecklistItem, EvalDimension
@@ -29,7 +28,7 @@ class JudgeRunner:
         system_prompt: str,
         checklist_items: list[ChecklistItem],
         eval_dimensions: list[EvalDimension],
-        pass_threshold: Decimal,
+        pass_threshold: float,
         temperature: float | None = None,
         max_tokens: int | None = None,
     ):
@@ -167,7 +166,7 @@ class JudgeRunner:
         # Evaluation average >= threshold
         if eval_scores:
             avg = sum(s["score"] for s in eval_scores) / len(eval_scores)
-            if avg < float(self.pass_threshold):
+            if avg < self.pass_threshold:
                 return False
 
         return True
